refactor: Move the salsa database to its own module
